// frontend/pizzaiolo/src/app/model/pizzaiolo/pizzasEditables.ts
/**
 * Api Documentation
 * Api Documentation
 *
 * OpenAPI spec version: 1.0
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
import { PizzasCortas } from './pizzasCortas';

/**
 * Version editable de las pizzas.
 */
export interface PizzasEditables { 
    /**
     * Pizza activa.
     */
    active?: boolean;
    /**
     * Coste total de la pizza.
     */
    amount?: number;
    /**
     * Descripción de la pizza.
     */
    description?: string;
    /**
     * Identificador de la base.
     */
    idBase?: number;
    /**
     * Identificador de la pizza.
     */
    idPizza?: number;
    /**
     * Identificador de la salsa.
     */
    idSauce?: number;
    /**
     * Ingredientes.
     */
    ingredientPizza?: Array<PizzasCortas>;
    /**
     * Precio neto de la pizza.
     */
    netPrice?: number;
}